# Conflicts:
#	src/MEGASync/stalled_issues/gui/stalled_issues_cases/StalledIssuesCaseHeaders.cpp
#	src/MEGASync/stalled_issues/gui/stalled_issues_cases/StalledIssuesCaseHeaders.h

+++ src/MEGASync/gui/node_selector/model/NodeSelectorModelItem.h
@@ -70,7 +70,7 @@ protected:
     QString mOwnerEmail;
     Status mStatus;
     bool mRequestingChildren;
-    long long mChildrenCounter;
+    int mChildrenCounter;
     bool mShowFiles;
     bool mChildrenAreInit;
